pyproject.toml
[tool.poetry]
name = "regrader"
version = "0.1.0"
description = "Online editor and zyBooks replacement"
authors = ["John Zlotek <jzlotek@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.1"
loguru = "^0.4.1"
python-dotenv = "^0.11.0"

[tool.poetry.dev-dependencies]
jedi = "^0.16.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
